fix portion (float) in done mqtt

+++ I4-QM-app/I4-QM-app/Models/Additive.cs
@@ -6,7 +6,7 @@
 
         public string Name { get; set; }
 
-        public int Portion { get; set; }
+        public float Portion { get; set; }
 
         public int Amount { get; set; }
